refactor(gui): simplify persistence component
